% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInterval.R
\name{computeInterval}
\alias{computeInterval}
\title{Computes statistics between first and last dates}
\usage{
computeInterval(data, dateColumn = "ObservationDate")
}
\arguments{
\item{data}{A data frame with a PatientMRN and date column}

\item{dateColumn}{A character vector giving the name of the date column}
}
\value{
Returns a data frame with the first and last date along with
the time between, number, and mean number of days between dates.
}
\description{
Computes the first and last date along with the time between, number, and mean number of days between dates.
}
