% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTherapyDuration.R
\name{computeTherapyDuration}
\alias{computeTherapyDuration}
\title{Compute duration per line of therapy}
\usage{
computeTherapyDuration(medicalTherapy)
}
\arguments{
\item{medicalTherapy}{A data frame of medical therapy from the Caisis database.}
}
\value{
Returns a data frame with medical therapy from the Caisis database with
the addition of line and duration of therapy.  A line of therapy is defined as drugs
initiated within 14 days of each other.  The duration is defined as the time in
days between two lines of therapy.
}
\description{
Adds duration per line of therapy to medical therapy extracted from the Caisis database.
}
