% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTreatmentDuration.R
\name{computeTreatmentDuration}
\alias{computeTreatmentDuration}
\title{Computes treatment duration per line of therapy}
\usage{
computeTreatmentDuration(data, days = 14)
}
\arguments{
\item{data}{A data frame generated by the function extractMedicalTherapy.}

\item{days}{Number of days between the start of treatment that defines a line of therapy.}
}
\value{
Returns a data frame with treatment duration calculated per line of therapy.  Output
contains the same input with the addition of 4 new columns where line sequentional number of
therapies, duration is the number of days between treatments, DaysFromTxStart is the number
of days from the first line of therapy, and MedTxDayByLine it the recomputed number of days
from the first line of therapy.
}
\description{
Computes treatment duration per line of therapy where a line of therapy as all drugs
given within a defined range of days.
}
