% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDiagnoses.R
\name{extractDiagnoses}
\alias{extractDiagnoses}
\title{Extract diagnoses}
\usage{
extractDiagnoses(connection, diagnoses = NULL, patients = NULL,
  format = "raw", n = -1)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{diagnoses}{A character vector of ICD9 or ICD10 codes.  If no limit is desired
then set as NULL. May use \% or _ for partial matches.}

\item{patients}{A character vector of patient medical record numbers.  If no limit is desired
then set as NULL.}

\item{format}{A character vector indicating the output format.  Options
include "raw" or "byDiagnosis".}

\item{n}{Number of records to retrieve.  Use n = -1 to retrieve all records.}
}
\value{
Returns a data frame with patient diagnoses and the date they were assigned in
the Diamonds database.
}
\description{
Extracts patient diagnoses from the Diamonds database and returns a data frame.
}
