% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDiamonds.R
\name{extractDiamonds}
\alias{extractDiamonds}
\title{Extract diamonds}
\usage{
extractDiamonds(channel, demographics = TRUE, observations, patients,
  format = "raw")
}
\arguments{
\item{channel}{RODBC connection object returned by odbcConnect.}

\item{demographics}{Boolean value indicating whether demographic information
should be included.}

\item{observations}{A character vector of observation identifiers. Refer to
the University of Washington Lab Medicine website to search of test codes
(\url{http://menu.labmed.washington.edu/oltg}).}

\item{patients}{A character vector of patient medical record numbers.}

\item{format}{A character vector indicating the output format.  Options
include "raw", "byObservationId", "byDaysFromFirstObservation", or
"byObservationDate".}
}
\value{
Returns a data frame with patient observations from the Diamonds database.
}
\description{
Extracts patient observations from the Diamonds database and returns a
data frame.
}
