% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEncounters.R
\name{extractEncounters}
\alias{extractEncounters}
\title{Extract clinical encounters}
\usage{
extractEncounters(connection, patients = NULL, n = -1)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{n}{Number of records to retrieve.  Use n = -1 to retrieve all records.}
}
\value{
Returns a data frame of clinical encounters from the Caisis database.
}
\description{
Extracts clinical encounters from the Caisis database and returns a data frame.
}
