% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLabs.R
\name{extractLabs}
\alias{extractLabs}
\title{Extract labs}
\usage{
extractLabs(connection, demographics = TRUE, labs = NULL, patients = NULL,
  format = "raw", n = Inf)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{demographics}{Boolean value indicating whether demographic information
should be included.}

\item{labs}{A character vector of lab test identifiers. Refer to
the University of Washington Lab Medicine website to search of test codes
(\url{http://menu.labmed.washington.edu/oltg}).  If no limit is desired
then set as NULL.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{format}{A character vector indicating the output format.  Options
include "raw", "byObservationId", "byDaysFromFirstObservation", or
"byObservationDate".}

\item{n}{Number of records to retrieve.  Use n = Inf to retrieve all records.}
}
\value{
Returns a data frame with patient labs from the Diamonds database.
}
\description{
Extracts patient labs from the Diamonds database and returns a data frame.
}
