% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMedicalTherapy.R
\name{extractMedicalTherapy}
\alias{extractMedicalTherapy}
\title{Extract medical therapy}
\usage{
extractMedicalTherapy(connection, patients = NULL, n = -1,
  duration = FALSE)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{n}{Number of records to retrieve.  Use n = -1 to retrieve all records.}

\item{duration}{A boolean indicating if the duration per line of therapy should
be computed.  A line of therapy is defined as drugs initiated within 14 days of each
other.  The duration is defined as the time in days between two lines of therapy.
If TRUE, then line and duraiton columns are added to the output.}
}
\value{
Returns a data frame with medical therapy from the Caisis database.
}
\description{
Extracts disease status from the Caisis database and returns a data frame.
}
