% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPathology.R
\name{extractPathology}
\alias{extractPathology}
\title{Extract pathology reports}
\usage{
extractPathology(connection, patients = NULL, n = -1)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{n}{Number of records to retrieve.  Use n = -1 to retrieve all records.}
}
\value{
Returns a data frame with pathology reports from the Caisis database.
}
\description{
Extracts pathology reports from the Caisis database and returns a data frame.
}
\details{
For columns reporting the precentatage of plasma cells, a new column
is created with the numeric value only.  Less or greater than symbols are
removed and the median is reported for ranges.  N/As are converted to 0.
}
