% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exractRadiationTherapy.R
\name{extractRadiationTherapy}
\alias{extractRadiationTherapy}
\title{Extract radiation therapy}
\usage{
extractRadiationTherapy(connection, patients = NULL, n = -1)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{n}{Number of records to retrieve.  Use n = -1 to retrieve all records.}
}
\value{
Returns a data frame with radiation therapy from the Caisis database.
}
\description{
Extracts radiation therapy from the Caisis database and returns a data frame.
}
