% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSurvival.R
\name{extractSurvival}
\alias{extractSurvival}
\title{Extract surivival information}
\usage{
extractSurvival(connection, patients = NULL, diagnosis = "Multiple Myeloma",
  group = "Plasma Cell Neoplasm")
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{patients}{A character vector of patient medical record numbers.  If no
limit is desired then set as NULL.}

\item{diagnosis}{Caisis diagnosis from which to extract the date of diagnoisis.  Survival
length in months will be computed from this date of diagnosis.}

\item{group}{Disease group from which to extract the last known alive date.}
}
\value{
Returns a data frame with PatientMRN, date of birth, last known alive date, last lab date,
and surival status and death dates from both Caisis and Diamonds databases.
From this information, the most recent date ('LastDate') is determined and the survival in
months is computed between this date and the date of diagnosis ('SurvivalMonths').
The surival status is the must up-to-date status from both Caisis and Diamonds databses.
The patient is assumed alive if the status is not dead and the CaisisLastAliveDate or
the LastLabDate is within the past 6 months.
}
\description{
Extracts surival information from the Caisis and Diamonds databases and returns a data frame.
}
