% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDiamonds.R
\name{filterDiamonds}
\alias{filterDiamonds}
\title{Filter diamonds}
\usage{
filterDiamonds(data, timepoints, range = "within", within = c(-14, 14),
  format = "raw")
}
\arguments{
\item{data}{A data frame extracted from the Diamonds database in "raw" format.}

\item{timepoints}{A data frame with the first column being the patient
medical record numbers and the second being the desired dates to filter
around (e.g. date of diagnosis or date of treatment).  There may be multiple
dates per patient.}

\item{range}{A character vector specifying the range of dates for each time
point to be extracted.  Available options include "on", "before", "after",
and  "within".  If "within" is selected a range of values must be provided in
the "within" argument.}

\item{within}{An integer vector of length 2 providing the number of dates
before and after each time point to filter.  For example, c(-14, 14) would
indicate that all dates less than or greater than 14 days of the specified
time point are to be filtered.}

\item{format}{A character vector indicating the output format.  Options
include "raw", "byObservationId", "byDaysFromFirstTimePoint", or
"byObservationDate".}
}
\value{
Returns a data frame with patient observations from the Diamonds
filtered by the desired time points database.  A new column is added giving
the number of days from the first specified time point.
}
\description{
Filters extracted Diamonds data by specified time points.
}
