% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePlasmacells.R
\name{parsePlasmacells}
\alias{parsePlasmacells}
\title{Parse pathology reports for \% plasma cells}
\usage{
parsePlasmacells(data)
}
\arguments{
\item{data}{A data frame with columns named "PatientMRN", "ReportId",
"PathNotes", and "ObservationDate".}
}
\value{
Returns a data frame with extracted \% plasma cells from pathology reports
}
\description{
Parses pathology reports for \% plasma cells
}
\details{
\describe{
\item{pattern1}{"\\\\d+\\.*\\\\d*\%plasmacells[^:0-9]"}
\item{pattern2}{"plasmacells:\\\\d+\\\\.*\\\\d*\%|plasmacells\\\\d+\\\\.*\\\\d*\%"}
\item{pattern3}{".\{50\}CD138.\{50\}|.\{50\}plasma.\{50\}|.\{50\}abnormal.\{50\}"}
}
}
