% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiamonds.R
\name{plotDiamonds}
\alias{plotDiamonds}
\title{Plot diamonds data}
\usage{
plotDiamonds(data, chart, by, interactive = FALSE)
}
\arguments{
\item{data}{A data frame extracted from the Diamonds database in "raw" format.}

\item{chart}{A character vector of the chart type.  Available options include
"histogram", "density", "box", "line".}

\item{by}{A character vector indicating whether the plots should be iterated
by patient or observation.  Available options include "PatientMRN" and
"ObservationId".}

\item{interactive}{A Boolean indicating if the plot should interactive.  If
TRUE then plot is interactive, if FALSE then the plot is static.}
}
\value{
Returns a histogram, density plot, blox plot, or line plot in ggplot
sytle.
}
\description{
Plots extracted data from the Diamonds database as a histogram, density plot,
blox plot, or line plot.
}
