\name{diamonds}
\alias{diamonds}
\alias{subdiamonds}
\alias{hexlines}
\alias{hiercenters}
\alias{subcenters}
\alias{hierlabels}
\alias{deep}
\alias{labeldeep}
\alias{centerdeep}
\alias{randlevel}
\alias{ranptdia}
\title{Internal functions for diamonds package}
\description{
  Various functions for internal use in the diamonds package.
}
\usage{
  diamonds(b, d)
  subdiamonds (b)
  hexlines (b, d)
  hiercenters (b, d)
  subcenters (b) 
  hierlabels (parent, level, d)
  deep (b, level, d)
  labeldeep (dt, parent, level, d)
  centerdeep (b, level, d)
  randlevel (tl, level, d)
  ranptdia (b)
}
\details{See R codes.}
\author{Denis White, \email{white.denis@epa.gov} }
\seealso{\code{\link{diamond.base}}}
\keyword{dplot}
