% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDiagnoses.R
\name{extractDiagnoses}
\alias{extractDiagnoses}
\title{Extract diagnoses}
\usage{
extractDiagnoses(connection, demographics = TRUE, diagnoses = NULL,
  patients = NULL, format = "raw", n = Inf)
}
\arguments{
\item{connection}{ODBC connection object returned by dbConnect.}

\item{demographics}{Boolean value indicating whether demographic information
should be included.}

\item{diagnoses}{A character vector of ICD9 or ICD10 codes.  If no limit is desired
then set as NULL.}

\item{patients}{A character vector of patient medical record numbers.  If no limit is desired
then set as NULL.}

\item{format}{A character vector indicating the output format.  Options
include "raw" or "byDiagnosis".}

\item{n}{Number of records to retrieve.  Use n = Inf to retrieve all records.}
}
\value{
Returns a data frame with patient diagnoses from the Diamonds database.
}
\description{
Extracts patient diagnoses from the Diamonds database and returns a data frame.
}
