% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKaryotype.R
\name{parseKaryotype}
\alias{parseKaryotype}
\title{Parse FISH karyotype strings}
\usage{
parseKaryotype(data, probes)
}
\arguments{
\item{data}{A data frame with columns named "PatientMRN", "ReportId",
"Source", "ObservationDate", and "Karyotype".}

\item{probes}{A data frame with columns named "probe" and "chr" that defines
the gene name and chromosomal location of the gene (e.g. 17p)}
}
\value{
Returns a data frame where each row represents a clone from the
FISH karyotype string.
}
\description{
Parses FISH karyotype strings into individual clones.
}
