% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePlasmacells.R
\name{parsePlasmacells}
\alias{parsePlasmacells}
\title{Parse pathology reports for \% plasma cells}
\usage{
parsePlasmacells(data)
}
\arguments{
\item{data}{A data frame with columns named "PatientMRN", "PathNum",
"PathNotes", and "ProcDate".}
}
\value{
Returns a data frame with extracted \% plasma cells from pathology reports
}
\description{
Parses pathology reports for \% plasma cells
}
\details{
\describe{
\item{pattern1}{"\\\\d+\\.*\\\\d*\%plasmacells[^:0-9]"}
\item{pattern2}{"plasmacells:\\\\d+\\\\.*\\\\d*\%|plasmacells\\\\d+\\\\.*\\\\d*\%"}
\item{pattern3}{".\{1,50\}CD138.\{1,50\}|.\{1,50\}plasma.\{1,50\}|.\{1,50\}abnormal.\{1,50\}"}
}
}
